reorganize code for helping to create data models for file and dir
