+++ src-tauri/src/filesystem/models/mod.rs
@@ -6,3 +6,5 @@ mod volume;
 pub use volume::VolumeInformation;
 mod directory_entries;
 pub use directory_entries::Entries;
+pub use directory_entries::{get_access_permission_number, get_access_permission_string, get_directory_size_in_bytes, count_subfiles_and_directories, format_system_time};
+
